train/pyproject.toml
[project]
name = "train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "datasets>=3.6.0",
    "torchaudio>=2.7.0",
    "transformers>=4.52.4",
    "unsloth>=2025.6.1",
]
